LPS-76949 portal-configuration-settings-test - adds common utility methods to the base test case
